+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/image/ImageOutput.java
@@ -23,6 +23,20 @@ import androidx.media3.common.util.UnstableApi;
 @UnstableApi
 public interface ImageOutput {
 
+  /** A no-op implementation of ImageOutput. */
+  ImageOutput NO_OP =
+      new ImageOutput() {
+        @Override
+        public void onImageAvailable(long presentationTimeUs, Bitmap bitmap) {
+          // Do nothing.
+        }
+
+        @Override
+        public void onDisabled() {
+          // Do nothing.
+        }
+      };
+
   /**
    * Called when an there is a new image available.
    *